split pet and illustratedGuide, enhence pet following
