+++ Assembly-CSharp.csproj
@@ -67,7 +67,7 @@
     <Compile Include="Assets\Scripts\PlayerSetting.cs" />
     <Compile Include="Assets\TutorialInfo\Scripts\Readme.cs" />
     <Compile Include="Assets\Scripts\TreasureHuntManager.cs" />
-    <Compile Include="Assets\following_user.cs" />
+    <Compile Include="Assets\Scripts\following_user.cs" />
     <Compile Include="Assets\Scripts\TreasureInspect.cs" />
     <Compile Include="Assets\Scripts\ButtonHandler.cs" />
     <Compile Include="Assets\Scripts\IllustratedGuide.cs" />
